!feat: Support multiple completions in backends
Generate multiple completions for a single prompt by
adding the num_return_sequences parameter.

This has limited support by the providers. The parameter for
the providers can have different names and may require a 2nd
parameter is also used.

The changing of results returned to be a list is a breaking change.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>