Use ModelDirectInputOutputProcessor as parent class directly
Following feedback in https://github.com/ibm-granite/granite-io/pull/83#issuecomment-2741229360,
it is suggested to creatye asubclass of ModelDirectInputOutputProcessor
directly instead of the wrapper function to return subclass to user.

This updates the example to show how to do this. It also updates the package paths
to enable direct access to relevant classes.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>